Release version 0.0.0
This version sends a text with the system-wide CPU and system-wide memory
usage every 5 seconds to a serial port receiver.

For sleeping the POSIX function sleep is used. This should be replaced by
the thrd_sleep function specified by the C11 standard once it becomes
available on the target platforms.